modify trafgen to track TX frames timestamps
This is a first draft which modify trafgen to generate timestamps for
each sending frame.
Timestamps are stored in the /tmp/out file with the following format:
* id: the 4 randomized bytes
* hardware timestamps s: this timestamps is generated by the NIC
* hardware timestamps ns: this timestamps is generated by the NIC
* software timestamps before sendto s
* software timestamps before sendto ns
* software timestamps after sendto s
* software timestamps after sendto ns

Each value is separated by a space.

To work, the NIC must support hardware timestamping and TX hardware
timestamping must be enabled. See hwstamp_ctl command of ptp4linux to
know if your NIC support  hardware timestamping and to enable TX.

SFL: #8529
Change-Id: Ifd1a15cbe0332a6ea4ea83db2dfbd9badf466e1d
Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>